backend/src/scripts/python/requirements.txt
# MLB Stats API Wrapper
MLB-StatsAPI==1.9.0

# Supabase Python Client
supabase==2.10.0

# Environment variable management
python-dotenv==1.0.0

# Date/time utilities (included in standard library, but explicit for clarity)
# datetime - standard library
# pytz - for timezone handling
pytz==2024.1
